cleanup in interactors
