+++ interactor/src/mvx_esdt_safe/mvx_esdt_safe_interactor_main.rs
@@ -12,9 +12,7 @@ use structs::{configs::EsdtSafeConfig, generate_hash::GenerateHash};
 use common_interactor::interactor_config::Config;
 use common_interactor::interactor_state::State;
 
-use common_test_setup::constants::{
-    INTERACTOR_WORKING_DIR, MVX_ESDT_SAFE_CODE_PATH, ONE_THOUSAND_TOKENS, SHARD_0,
-};
+use common_test_setup::constants::{INTERACTOR_WORKING_DIR, MVX_ESDT_SAFE_CODE_PATH, SHARD_0};
 
 pub struct MvxEsdtSafeInteract {
     pub interactor: Interactor,
@@ -87,16 +85,7 @@ impl MvxEsdtSafeInteract {
         ];
 
         for (ticker, token_type, decimals) in token_configs {
-            let amount = if matches!(
-                token_type,
-                EsdtTokenType::NonFungibleV2 | EsdtTokenType::DynamicNFT
-            ) {
-                BigUint::from(1u64)
-            } else {
-                BigUint::from(ONE_THOUSAND_TOKENS)
-            };
-
-            self.create_token_with_config(token_type, ticker, amount, decimals)
+            self.create_token_with_config(token_type, ticker, decimals)
                 .await;
         }
     }
